+++ dx2syn/rates_test.go
@@ -6,7 +6,7 @@ import (
 
 func TestComputeDurationsMs(t *testing.T) {
 	expect := [4]int{1478, 2, 0, 54312}
-	actual := computeDurationsMs([4]int{99, 50, 40, 0}, [4]int{30, 99, 99, 10})
+	actual := computeDurationsMs([4]byte{99, 50, 40, 0}, [4]byte{30, 99, 99, 10})
 	for i, exp := range expect {
 		if actual[i] != exp {
 			t.Errorf("ms[%d] expected %d, got '%d'\n", i, exp, actual[i])
@@ -17,7 +17,7 @@ func TestComputeDurationsMs(t *testing.T) {
 
 func TestComputeDurationsMs1(t *testing.T) {
 	expect := [4]int{1478, 2, 0, 54312}
-	actual := computeDurationsMs([4]int{99, 50, 40, 0}, [4]int{30, 99, 99, 10})
+	actual := computeDurationsMs([4]byte{99, 50, 40, 0}, [4]byte{30, 99, 99, 10})
 	for i, exp := range expect {
 		if actual[i] != exp {
 			t.Errorf("ms[%d] expected %d, got '%d'\n", i, exp, actual[i])
@@ -28,7 +28,7 @@ func TestComputeDurationsMs1(t *testing.T) {
 
 func TestComputeDurationsMs2(t *testing.T) {
 	expect := [4]int{2, 42, 13, 845}
-	actual := computeDurationsMs([4]int{99, 95, 60, 0}, [4]int{90, 50, 80, 50})
+	actual := computeDurationsMs([4]byte{99, 95, 60, 0}, [4]byte{90, 50, 80, 50})
 	for i, exp := range expect {
 		if actual[i] != exp {
 			t.Errorf("ms[%d] expected %d, got '%d'\n", i, exp, actual[i])
